Move ProjectData ↔ StringResources functions
